tests/wpt/web-platform-tests/feature-policy/permissions-policy-feature-policy-coexist.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src=/resources/testharness.js> </script>
<script src=/resources/testharnessreport.js> </script>
<div>
  This is a page with following headers: <br>
  <div>
    Feature-Policy: geolocation 'none', fullscreen 'none'<br>
    Permissions-Policy: geolocation=self, payment=()
  </div>
</div>
<script>
  const policy = document.featurePolicy;
  test(() => {
    assert_true(policy.allowsFeature('geolocation'));
  }, "When there is conflict in Feature Policy header and Permissions Policy" +
  "header, Permissions Policy wins.");

  test(() => {
    assert_false(policy.allowsFeature('fullscreen'));
    assert_false(policy.allowsFeature('payment'));
  }, "When there is no conflict, Feature Policy and Permissions Policy should " +
  "all be able to control each feature by themselves.");
</script>
